Add files via upload
For this Video Game assignment, I programmed the Hangman Game, The game rules are coded to follow the traditional game. Input one letter at a time, if the character is not a part of the answer, then you will lose, and if the character is a part of the final answer, you will win. The amount of times you can input a letter is a fixed number, based on there only being one answer for this program. The game is programmed so the user can play an infinite amount of times and updates how the hangman looks based on your input being correct or incorrect.